CVC_YOLOv3/requirements.txt
#python 3.7

numpy==1.16.4
matplotlib==3.1.0
torchvision==0.3.0
opencv_python==4.1.0.25
torch==1.1.0
requests==2.20.0
pandas==0.24.2
imgaug==0.3.0
onnx==1.6.0
optuna==0.19.0
Pillow==6.2.1
protobuf==3.11.0
pymysql==0.9.3
retrying==1.3.3
tensorboardX==1.9
tqdm==4.39.0
